Rename e2e test files
