+++ e2e/nx-firebase-e2e/test-utils/index.ts
@@ -1,7 +1,8 @@
-export * from './test-helpers'
-export * from './test-logger'
-export * from './test-imports'
-export * from './test-commands'
-export * from './test-project-data'
-export * from './test-apps'
-export * from './test-functions'
+export * from './test-utils-apps'
+export * from './test-utils-commands'
+export * from './test-utils-functions'
+export * from './test-utils-helpers'
+export * from './test-utils-imports'
+export * from './test-utils-logger'
+export * from './test-utils-project-data'
+
