+++ e2e/nx-firebase-e2e/test-utils/test-utils-imports.ts
@@ -1,4 +1,4 @@
-import type { ProjectData } from './test-project-data'
+import type { ProjectData } from './test-utils-project-data'
